fix: evaluate TYPE_CHECKING guards + lint fixes
- Audited 6 TYPE_CHECKING imports, removed unused gate import
- Fixed SIM117 nested with statements in security tests
- Fixed import sorting
