+++ silas/models/work.py
@@ -4,16 +4,15 @@ import json
 from datetime import datetime
 from enum import StrEnum
 from hashlib import sha256
-from typing import TYPE_CHECKING, Literal
+from typing import Literal
 
 from pydantic import BaseModel, Field, field_validator, model_validator
 
 from silas.models.agents import InteractionMode
+from silas.models.approval import ApprovalToken
+from silas.models.expectation import Expectation
+from silas.models.gates import AccessLevel, Gate
 from silas.models.messages import utc_now
-
-if TYPE_CHECKING:
-    from silas.models.approval import ApprovalToken
-    from silas.models.gates import AccessLevel, Gate
 
 
 class WorkItemType(StrEnum):
@@ -65,34 +64,6 @@ class BudgetUsed(BaseModel):
         self.attempts += child.attempts
         self.planner_calls += child.planner_calls
         self.executor_runs += child.executor_runs
-        return self
-
-
-class Expectation(BaseModel):
-    exit_code: int | None = None
-    equals: str | None = None
-    contains: str | None = None
-    regex: str | None = None
-    output_lt: float | None = None
-    output_gt: float | None = None
-    file_exists: str | None = None
-    not_empty: bool | None = None
-
-    @model_validator(mode="after")
-    def _validate_exactly_one_predicate(self) -> Expectation:
-        checks = [
-            self.exit_code is not None,
-            self.equals is not None,
-            self.contains is not None,
-            self.regex is not None,
-            self.output_lt is not None,
-            self.output_gt is not None,
-            self.file_exists is not None,
-            self.not_empty is True,
-        ]
-        selected = sum(checks)
-        if selected != 1:
-            raise ValueError("Expectation must define exactly one predicate")
         return self
